requirements.txt
google-generativeai
groq
pydantic
python-dotenv
requests
